tests(cli): add some more
